+++ server/lib/src/server/migrations.rs
@@ -615,13 +615,7 @@ impl<'a> QueryServerWriteTransaction<'a> {
 
         // Check the admin object exists (migrations).
         // Create the default idm_admin group.
-        let admin_entries: Vec<EntryInitNew> = vec![
-            BUILTIN_ACCOUNT_ANONYMOUS_V1.clone().into(),
-            BUILTIN_ACCOUNT_ADMIN.clone().into(),
-            BUILTIN_ACCOUNT_IDM_ADMIN.clone().into(),
-            BUILTIN_GROUP_IDM_ADMINS_V1.clone().try_into()?,
-            BUILTIN_GROUP_SYSTEM_ADMINS_V1.clone().try_into()?,
-        ];
+        let admin_entries: Vec<EntryInitNew> = idm_builtin_admin_entries()?;
         let res: Result<(), _> = admin_entries
             .into_iter()
             // Each item individually logs it's result
@@ -632,49 +626,7 @@ impl<'a> QueryServerWriteTransaction<'a> {
         debug_assert!(res.is_ok());
         res?;
 
-        // Create any system default schema entries.
-        let idm_entries: Vec<&BuiltinGroup> = vec![
-            &IDM_ALL_PERSONS,
-            &IDM_ALL_ACCOUNTS,
-            &IDM_PEOPLE_MANAGE_PRIV_V1,
-            &IDM_PEOPLE_ACCOUNT_PASSWORD_IMPORT_PRIV_V1,
-            &IDM_PEOPLE_EXTEND_PRIV_V1,
-            &IDM_PEOPLE_SELF_WRITE_MAIL_PRIV_V1,
-            &IDM_PEOPLE_WRITE_PRIV_V1,
-            &IDM_PEOPLE_READ_PRIV_V1,
-            &IDM_HP_PEOPLE_EXTEND_PRIV_V1,
-            &IDM_HP_PEOPLE_WRITE_PRIV_V1,
-            &IDM_HP_PEOPLE_READ_PRIV_V1,
-            &IDM_GROUP_MANAGE_PRIV_V1,
-            &IDM_GROUP_WRITE_PRIV_V1,
-            &IDM_GROUP_UNIX_EXTEND_PRIV_V1,
-            &IDM_ACCOUNT_MANAGE_PRIV_V1,
-            &IDM_ACCOUNT_WRITE_PRIV_V1,
-            &IDM_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            &IDM_ACCOUNT_READ_PRIV_V1,
-            &IDM_RADIUS_SECRET_WRITE_PRIV_V1,
-            &IDM_RADIUS_SECRET_READ_PRIV_V1,
-            &IDM_RADIUS_SERVERS_V1,
-            // Write deps on read, so write must be added first.
-            &IDM_HP_ACCOUNT_MANAGE_PRIV_V1,
-            &IDM_HP_ACCOUNT_WRITE_PRIV_V1,
-            &IDM_HP_ACCOUNT_READ_PRIV_V1,
-            &IDM_HP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            &IDM_SCHEMA_MANAGE_PRIV_V1,
-            &IDM_HP_GROUP_MANAGE_PRIV_V1,
-            &IDM_HP_GROUP_WRITE_PRIV_V1,
-            &IDM_HP_GROUP_UNIX_EXTEND_PRIV_V1,
-            &IDM_ACP_MANAGE_PRIV_V1,
-            &DOMAIN_ADMINS,
-            &IDM_HP_OAUTH2_MANAGE_PRIV_V1,
-            &IDM_HP_SERVICE_ACCOUNT_INTO_PERSON_MIGRATE_PRIV,
-            &IDM_HP_SYNC_ACCOUNT_MANAGE_PRIV,
-            // All members must exist before we write HP
-            &IDM_HIGH_PRIVILEGE_V1,
-            // other things
-            &IDM_UI_ENABLE_EXPERIMENTAL_FEATURES,
-            &IDM_ACCOUNT_MAIL_READ_PRIV,
-        ];
+        let idm_entries = idm_builtin_non_admin_groups();
 
         let res: Result<(), _> = idm_entries
             .into_iter()
