+++ server/testkit/Cargo.toml
@@ -39,6 +39,8 @@ tracing = { workspace = true, features = ["attributes"] }
 tokio = { workspace = true, features = ["net", "sync", "io-util", "macros"] }
 openssl = { workspace = true }
 lazy_static = { workspace = true }
+petgraph = { version = "0.6.4", features = ["serde"] }
+regex.workspace = true
 
 
 [build-dependencies]
