+++ tools/cli/Cargo.toml
@@ -3,7 +3,6 @@ name = "kanidm_tools"
 default-run = "kanidm"
 description = "Kanidm Client Tools"
 documentation = "https://kanidm.github.io/kanidm/stable/"
-
 version = { workspace = true }
 authors = { workspace = true }
 rust-version = { workspace = true }
@@ -52,7 +51,8 @@ url = { workspace = true, features = ["serde"] }
 uuid = { workspace = true }
 zxcvbn = { workspace = true }
 lazy_static.workspace = true
-regex.workspace = true
+regex = { workspace = true }
+sketching = { workspace = true }
 
 [dependencies.cursive]
 version = "0.20.0"
